Continue prototyping trying to have different views :books:
